msgpack: WIP decoder, works but the code needs to be cleaned up
There are parts of the code that weren't looped because since there was no key wrapping the headers that meant that the values element had to be processed in the same block which was nasty.

Now that the functionality is there I'll make it pretty and implement the necessary cleanup code for error conditions.