prometheus_remote_write: fixed a bug where the sample count was set to a higher number than the samples set
This is just for the sake of leaving it at a point where it works. I think the way in which this is made is wrong because each metric entry maps to a one element timeseries.

Because of this on the next commit I will update the protobuf definitions to be able to use Samples and Excemplars to put all the metric values in a single the series with the appropriate data.