tests: encoding: fixed a few tests that failed
There is a conflict between this PR and 102, PR 102
updates the Summary related metric field usage in the
decoders and encoders whilst this one fixes a leak in
the map destructor.

In this PR none of the tests that expect sum_quantiles to
be a fixed size array allocated in the cmt_metric structure
will pass because they don't know (yet) they have to allocate
that array and thus end up accessing invalid memory.

Signed-off-by: Leonardo Alminana <leonardo@calyptia.com>