prometheus_remote_write: Add cutoff for outdated metrics
This is because the official Prometheus node_exporter should remove the
outdated metrics because it creates the storages which should be
handling/storing metrics' information on every cycles.
At least, we should cut off for 1 hour or older metrics to prevent
metric too old errors on prometheus remote write mechanism.

Signed-off-by: Hiroshi Hatake <hatake@calyptia.com>