out_prometheus_remote_write: update cutoff treshold
This treshold was added to stop metrics older than one hour 
being sent to prometheus. It seems that during compilation, atleast on windows, 
this number gets changed to 817405952, which is significantly lower and 
results in all metrics being thrown.

The treshold have been changed to a static value that persists through compilation.

Signed-off-by: gmhomb <gjerthomb@hotmail.com>